libra_defconfig: Make Ion allocate in page-sized chunks
With this option disabled, Ion allocates memory in order-4 sized chunks
(16 pages at a time), which is quite a heavy burden on the page
allocator, and can take a while to allocate when memory is fragmented
after a lot of uptime. To lower the allocation latency in Ion, have it
only allocate in page-sized chunks.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mortezadadgar <mortezadadgar97@gmail.com>

Change-Id: Iea0c6a13f5d6d530419ef3b5df52e46e0b12c564